Pi Quadcopter
PiQuadcopter based Project, running on the Raspberry Pi. Note: that I've used the Raspberry Pi 3 but it was originally tested on a Raspberry Pi 2 according to Richard Wendel III (user rgw3d)
I would like to thank user rgw3d that inspired this one and give full credit to him.